Lab4-middleware-hw/I2-servants/client/requirements.txt
zeroc-ice>=3.7.0